TurtleCoin v1.0.1
That's it, it's capped off, there is no more.

A heartfelt thank you goes out to the community, the contributors, the developers, the late night warriors that have helped make this project great.

See you on the other side
  - RockSteady (RIP)

A special thanks to the following contributors who submitted code to this repo directly (in descending order of commits, changes, etc.)

@zpalmtree
@brandonlehmann
@RocksteadyTC
@rashedmyt
@mrrovot
@thinkpol2
@dirtybits
@rixombea
@nnamon
@jprjr
@wrkzdev
@sumlnoether
@SoreGums
@yamiM0NSTER
@andrewnk
@hensleyrob
@LeoStehlik
@ExtraHash
@funkypenguin
@gxhshw
@lamby
@wreiner
@bwentzloff
@crappyrules
@ngeojiajun
@blakelapierre
@DaveLong
@trashbandit
@deertacos
@holytastyguacamole
@michelangelo314
@scottjw
@imperdin
@mager
@pypper
@0des
@HashVault
@cryptoBOOOOM
@CutterDev
@krruzic
@ftwillms
@Watt3r
@dmaillard
@Swardu
@hooftly
@Sajo8
@mark9064
@CaptainMeatloaf
@devinbileck
@anonanonymous
@fexra
@DevAurelius
@astrorigin
@mlametta
@iwoojoo
@gregmeszaros
@SylarBend